prepare for upgrade to 5.1

+++ database/migrations/2015_06_03_151644_create_shop_products_table.php
@@ -21,7 +21,7 @@ class CreateShopProductsTable extends Migration {
             $table->text('description')->nullable();
             $table->string('weight')->nullable();
             $table->integer('sku');
-            $table->decimal('price', 6, 2);
+            $table->integer('price');
             $table->boolean('is_downloadable')->default(false);
             $table->tinyInteger('hidden');
             $table->softDeletes();
